Marching Squares and Perlin Noise.
This fun little project Uses Perlin Noise to procedurally produce shapes and marching squares to encapsulate them. A lot of it can be tweaked to make cool looking stuff. Might try and do a 3D implementation in Unity later. Some formula creation and logic are in the pictures listed in the files.


Cheers!